--- src/benchmark.txt
@@ -1,14 +1,0 @@
-
-Iterations: 50
-Samples: 100
-
-Compute Shader SSBO:
-Mean Time Per Sample(MTPS) = 4.05817e-05s
-Net Time(NT) = 62.8212s
-
-Compute Shader Texture:
-
-
-
-Fragment Shader:
-
