+++ src/splat_queue.h
@@ -1,6 +1,5 @@
-#ifndef SPLAT_HANDLER_H
-#define SPLAT_HANDLER_H
-
+#ifndef SPLAT_QUEUE_H
+#define SPLAT_QUEUE_H
 
 
 #include <queue>
@@ -8,102 +7,64 @@
 
 
 
+
 struct splat{
     vec2 i_pos, f_pos, dir; 
 };
 
+
 class SplatQueue{
 
 
-public: 
+    public: 
 
-    SplatQueue() {
+        SplatQueue();
 
-        cur_splat.i_pos[0] = -1;
-        cur_splat.i_pos[1] = -1;
+        SplatQueue(int sim_width, int sim_height);
 
-    }
 
-    void build_splat(double x, double y){
+        void build_splat(double x, double y);
 
-        x = std::min(std::max(0.0, x), double(win_width) - 1) * float(SIM_WIDTH) / float(win_width);
-        y = std::min(std::max(0.0, y), double(win_height) - 1)* float(SIM_HEIGHT) / float(win_height);
-        y = win_height - y;
+        void push(splat s);
 
-        if(!start_trail){
-            start_trail = true;
-            cur_splat.i_pos[0] = x;
-            cur_splat.i_pos[1] = y;
-        }else if(!(cur_splat.i_pos[0] == x && cur_splat.i_pos[1] == y)){
-            cur_splat.f_pos[0] = x;
-            cur_splat.f_pos[1] = y;
-            push(cur_splat);
-        }  
-    }
+        splat pop();
 
-    void push(splat s){
-        queue.push(s);
-        cur_splat.i_pos = cur_splat.f_pos;
-    }
+        bool isEmpty();
 
-    splat pop(){
+        int size();
+
+        void clear();
         
-        splat ret = queue.front();
-        ret.dir = (ret.f_pos - ret.i_pos);
-        ret.dir = ret.dir / ret.dir.length();
-        queue.pop();
-        return ret;
-    }
-
-    bool isEmpty(){
-        return queue.empty();
-    }
-
-    int size(){
-        return queue.size();
-    }
-
-    void clear(){
-        for(int i = 0; i < size(); i++){
-            queue.pop();
-        }
-    }
-
-    void end_trail(){
-        start_trail = false;
-    }
+        void end_trail();
     
-   
-public:
+    public:
 
-    void static resize(int new_win_width, int new_win_height){
-        win_width = new_win_width;
-        win_height = new_win_height;
-    }
+        void static resize(int new_win_width, int new_win_height){
+            win_width = new_win_width;
+            win_height = new_win_height;
+        }
 
  
 
     
 
-private:
-    bool start_trail = false;
-    splat cur_splat;
-    std::queue<splat> queue; 
-    
+    private:
+        bool start_trail = false;
+        splat cur_splat;
+        std::queue<splat> queue; 
+        
 
 
 
-private:
-    static const unsigned int SIM_WIDTH, SIM_HEIGHT;
-    static unsigned int win_width, win_height;
+    private:
+        const unsigned int SIM_WIDTH, SIM_HEIGHT;
+        static inline unsigned int win_width, win_height =0 ;
 
 
 };
 
-
-unsigned int SplatQueue::win_width = 0;
-unsigned int SplatQueue::win_height = 0;
-
+// unsigned int SplatQueue::win_width = 0;
+// unsigned int SplatQueue::win_height = 0;
 
 
 #endif
